refactor: test runner public API
